Manual Object Measurement/requirement.txt
streamlit
opencv-python
streamlit-webrtc
Pillow
numpy
streamlit-drawable-canvas
